Refactor AI methods from AiWebParser into environment adapters
Moves sendAiRequest, loadImageAsBase64, and OCR caching logic from
AiWebParser to ScriptableAdapter and WebAdapter. Updates AiWebParser
to use dependency injection for these services, ensuring strict
separation of concerns and environment-agnostic parsing logic.
